% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline names.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://cran.r-project.org/package=nycflights13, v0.2.2
}
\usage{
airlines
}
\description{
Look up airline names from their carrier codes.
}
\examples{
\dontrun{
  airlines <- nycflights13::airlines
  save(airlines, file = "data/airlines.rda", compress = "xz")
}

}
\keyword{datasets}
