% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference.R
\name{build_reference}
\alias{build_reference}
\title{Build and add the 'reference' table to the package}
\usage{
build_reference(package = NULL, quiet = FALSE)
}
\arguments{
\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{build_reference()} builds and adds the 'reference' entity table of
the Systematic Quantitative Literature Review (SQLR) system to an R package.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{?write_metadata()} to learn more.
}
\examples{
\dontrun{
build_reference()}
}
\seealso{
Other SQLR system functions: 
\code{\link{keyword_set}()},
\code{\link{normalize_extdata}()},
\code{\link{read_sheet}()},
\code{\link{start_sqlr}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{SQLR system functions}
