% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_dtm}
\alias{col_dtm}
\alias{col_dtm.MariaDBConnection}
\title{Generate SQL for a date/time data type definition}
\usage{
col_dtm(..., con = get_con())

\method{col_dtm}{MariaDBConnection}(type = c("datetime", "date", "time",
  "year"), val = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{type}{One of \code{datetime}, \code{date}, \code{time}, \code{year},
specifying the type of date/time of the column.}

\item{val}{A value, the class of which is used for automatically inferring
type.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for date/time data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
