% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_id}
\alias{col_id}
\alias{col_id.MariaDBConnection}
\title{Generate SQL for an id column}
\usage{
col_id(..., con = get_con())

\method{col_id}{MariaDBConnection}(name = "id", type = "int",
  unsigned = TRUE, auto_increment = TRUE, key = "primary", ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{name}{String specifying the column name}

\item{type}{Either a call to one of the functions capable of specifying a
column type (where the con object will be replaced with the one from the
enclosing function call), or the name of one of the functions (as string
or function object) with arguments specified in the enclosing function call.}

\item{unsigned}{Logical switch specifying whether the integer is singed or
unsigned.}

\item{auto_increment}{Logical switch specifying whether the column is auto
incrementing.}

\item{key}{One of \code{NA}, \code{unique}, \code{primary}, \code{key},
specifying if and what type of index the column represents.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for an id column specification
(typically an auto incrementing, unsigned integer type primary key column),
in column definitions for CREATE/ALTER TABLE statements.
}
