% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-spec.R
\name{col_raw}
\alias{col_raw}
\alias{col_raw.MariaDBConnection}
\title{Generate SQL for a binary data type definition}
\usage{
col_raw(..., con = get_con())

\method{col_raw}{MariaDBConnection}(length = 255L, fixed = FALSE,
  force_blob = length >= 16384L, ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{length}{The maximum number of bytes the column is expected to hold.}

\item{fixed}{Logical switch specifying whether lengths are fixed throughout
rows.}

\item{force_blob}{Length threshold for forcing the type from (VAR)BINARY to
BLOB which causes the values to be stored separately from the rest of the
row. BLOB columns only contribute 9 to 12 bytes toward the row size limit
(65,535 bytes).}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for binary data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
