% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_spec}
\alias{col_spec}
\title{Generate SQL for a column definition}
\usage{
col_spec(name = paste(sample(letters, 10, TRUE), collapse = ""),
  type = col_int(), nullable = TRUE, default = NULL,
  auto_increment = FALSE, key = c(NA, "unique", "primary", "key"),
  comment = NULL, ...)
}
\arguments{
\item{name}{String specifying the column name}

\item{type}{Either a call to one of the functions capable of specifying a
column type (where the con object will be replaced with the one from the
enclosing function call), or the name of one of the functions (as string
or function object) with arguments specified in the enclosing function call.}

\item{nullable}{Logical switch specifying whether the column may contain
NULL or not.}

\item{default}{String specifying the default value.}

\item{auto_increment}{Logical switch specifying whether the column is auto
incrementing.}

\item{key}{One of \code{NA}, \code{unique}, \code{primary}, \code{key},
specifying if and what type of index the column represents.}

\item{comment}{An optional comment string.}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for column definitions in
CREATE/ALTER TABLE statements.
}
\examples{
\dontrun{
col_spec(name = "foo", type = col_int())
col_spec(name = "foo", type = col_int(unsigned = TRUE))

col_spec(name = "foo", type = col_int)
col_spec(name = "foo", type = col_int, unsigned = TRUE)

col_spec(name = "foo", type = "integer")
col_spec(name = "foo", type = "integer", unsigned = TRUE)

col_spec(name = "foo", type = "int")
col_spec(name = "foo", type = "int", unsigned = TRUE)
}

}
