% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-con.R
\docType{data}
\name{config}
\alias{config}
\title{Config environment}
\format{An object of class \code{environment} of length 0.}
\usage{
config
}
\description{
Configuration information, as well as the current connection
is saved to and fetched from this environment.
}
\keyword{datasets}
