% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_db}
\alias{connect_db}
\title{Connect to a database}
\usage{
connect_db(config)
}
\arguments{
\item{dbType}{The type of database to connect to (default: "mysql")}
}
\value{
A connection object that can be used for further database access.
}
\description{
Depending on the type selected by the user, the appropriate
function is called to set up and/or connect to a database.
}
