% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_mysql}
\alias{connect_mysql}
\title{Connect to a MySQL database}
\usage{
connect_mysql(...)
}
\arguments{
\item{...}{Arguments passed to \code{dbConnect}}
}
\value{
Connection object that can be used for further database access. The
connection is automatically destroyed when garbage collected.
}
\description{
This function provides a wrapper around \code{dbConnect} that
sets up a database and user with all privileges to that schema if any of the
two does not already exist.
}
