% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ref.R
\name{count_ref}
\alias{count_ref}
\title{Count references/citations from files}
\usage{
count_ref(file = file.choose())
}
\arguments{
\item{file}{(optional) a \code{character} object indicating RIS or ZIP file names.
If not assigned, a dialog window will be open enabling the user to search
and select a file (only for interactive sessions).}
}
\value{
A \code{numeric} value with the count or sum of the count of the
references/citations found in the \code{file} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{count_ref()} count references/citations from files. If more than one file is
used, the function returns the sum of the count of the references/citations
from each file.

At the moment, \code{count_ref()} works only with PubMed and RIS (Research
Information Systems) formats.
}
\details{
\code{count_ref()} also allows you to read ZIP compacted files. Just assign
the ZIP file in the \code{file} argument.
}
\examples{
\dontrun{
references <- raw_data("reference")
file <- references[grep("_apa_", references)]
file <- raw_data("reference", file)

count_ref(file)}
}
\seealso{
Other reference/citation functions: 
\code{\link{guess_ref}()},
\code{\link{read_ref}()}
}
\concept{reference/citation functions}
