% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ris.R
\name{count_ris}
\alias{count_ris}
\title{Count citations/references from RIS files}
\usage{
count_ris(file = file.choose())
}
\arguments{
\item{file}{(optional) A \code{character} object indicating RIS or ZIP file names.
If not assigned, a dialog window will be open enabling the user to search
and select a file (only for interactive sessions).}
}
\value{
A \code{numeric} object with the count or sum of the count of the
citations/references found in the 'file' argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{count_ris()} count citations/references from RIS (Research Information
Systems) files. If more than one file is used, the function returns the sum
of the count of the citations/references from each file.
}
\details{
\code{count_ris()} also allows you to read zip compacted files. Just assign
the zip file in the \code{file} argument.
}
\examples{
\dontrun{
file <- "2021-04-28 - Citations - APA - EN - 1-1150.zip"
file <- raw_data("citation", file)

count_ris(file)}
}
\seealso{
Other citation functions: 
\code{\link{read_ris}()}
}
\concept{citation functions}
