% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutter.R
\name{cutter}
\alias{cutter}
\title{Cut a vector into pieces}
\usage{
cutter(x, index, between = NULL, rm_start = FALSE, rm_end = FALSE)
}
\arguments{
\item{x}{An \link[checkmate:checkAtomicVector]{atomic vector} (\emph{e.g.},
\code{character}, \code{integer}, \code{numeric}, \code{factor}, \code{POSIXct}).}

\item{index}{An \link[checkmate:checkIntegerish]{integerish} \code{numeric} object
or an \code{integer} object with the indexes/cutting points.}

\item{between}{(optional) A string object indicating the direction of the cut
(choices: \code{"left"}, \code{"right"}). This argument only need to be assigned if
the cut must be performed between the indexes values (default: \code{NULL}).}

\item{rm_start}{(optional) a \code{logical} value indicating if the start element
of the cut must be removed (default: \code{FALSE}).}

\item{rm_end}{(optional) a \code{logical} value indicating if the end element
of the cut must be removed (default: \code{FALSE}).}
}
\value{
A \code{list} object with the cut pieces as elements.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{cutter()} cut vector objects into pieces by cutting points/indexes.
}
\details{
\code{cutter()} can perform different kinds of cuts. Here are some examples.
\subsection{Cutting by index values}{\preformatted{cutter(seq(10), c(3, 9))

   cut         cut
    |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2
Element 2: 4, 5, 6, 7, 8
Element 3: 10
}
}

\subsection{Cutting between index values}{\preformatted{cutter(seq(10), c(3, 9), between = "left")

  cut         cut
   |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2
Element 2: 3, 4, 5, 6, 7, 8
Element 3: 9, 10
}\preformatted{cutter(seq(10), c(3, 9), between = "right")

    cut         cut
     |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2, 3
Element 2: 4, 5, 6, 7, 8, 9
Element 3: 10
}
}

\subsection{Removing tips}{\preformatted{cutter(seq(20), c(7, 16), rm_start = TRUE, rm_end = TRUE)

           cut                      cut
            |                        |
1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
|---------|                            |---------|
 start tip                               end tip

Element 1: 8, 9, 10, 11, 12, 13, 14, 15
}
}
}
\examples{
## Cutting by index values

cutter(seq(10), c(3, 9))

## Cutting between index values

cutter(seq(10), c(3, 9), between = "left")

cutter(seq(10), c(3, 9), between = "right")

## Removing start or end tips

cutter(seq(10), c(3, 9), rm_start = TRUE)

cutter(seq(10), c(3, 9), rm_end = TRUE)
}
\seealso{
Other utility functions: 
\code{\link{raw_data}()}
}
\concept{utility functions}
