% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-def.R
\name{drop_db_tbl}
\alias{drop_db_tbl}
\alias{drop_db_tbl.MariaDBConnection}
\title{Delete database table(s)}
\usage{
drop_db_tbl(..., con = get_con())

\method{drop_db_tbl}{MariaDBConnection}(tbls, temporary = FALSE,
  force = FALSE, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{tbls}{Character vector indicating which table(s) to delete.}

\item{temporary}{Logical switch, restricting the delete to temporary tables.}

\item{force}{Logical switch; if \code{TRUE}, no error is thrown if one of
the tables does not exist.}
}
\value{
A single logical value, indicating whether the statement completed
successfully (invisibly).
}
\description{
Remove one or more tables from database.
}
