% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-spec.R
\name{fk_spec}
\alias{fk_spec}
\title{Generate SQL for reference definition}
\usage{
fk_spec(..., con = get_con())
}
\arguments{
\item{child_ind}{Index column name(s) from which the reference is pointing
to.}

\item{parent_tbl}{The table to which the reference is pointing to.}

\item{parent_ind}{Index column name(s) to which the reference is pointing
to.}

\item{constr_name}{(Optional) name of the constraint.}

\item{fk_name}{(Optional) name of the foreign key.}

\item{match}{One of \code{NA}, \code{simple}, \code{full}, \code{partial}.
For MySQL, NA is recommended.}

\item{on_del, on_upd}{One of \code{restrict}, \code{cascade},
\code{set_null}, \code{no_action}, \code{set_default}, specifying the
behavior of \code{UPDATE} and \code{DELETE} operations which affect key
values in the parent table that have matching rows in the child table.}

\item{check}{Logical switch for checking whether parent table/columns are
available.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for foreign key reference
definitions e.g. in CREATE/ALTER TABLE statements.
}
\section{TODO}{
 improve checks and add check unit test as soon as tables can
be created.
}

