% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flights data}
\format{Data frame with columns
\describe{
\item{departure}{Date and time of departure as a \code{POSIXct} date}
\item{dep_delay,arr_delay}{Departure and arrival delays, as a
  \code{difftime} object. Negative times represent early
  departures/arrivals.}
\item{air_time}{Amount of time spent in the air, in minutes}
\item{carrier}{Two letter carrier abbreviation. See \code{\link{airlines}}
  to get name}
\item{flight}{Flight number}
\item{tailnum}{Plane tail number}
\item{origin,dest}{Origin and destination. See \code{\link{airports}} for
  additional metadata.}
\item{distance}{Distance between airports, in miles}
}}
\source{
https://cran.r-project.org/package=nycflights13, v0.2.2
}
\usage{
flights
}
\description{
On-time data for all flights that departed NYC (i.e. JFK, LGA or EWR) in
2013.
}
\examples{
\dontrun{
  flights <- nycflights13::flights
  flights <- flights[flights$month \%in\% 1:3, ]
  tmp <- sprintf("\%04d", flights$dep_time)
  tmp[is.na(flights$dep_time)] <- NA
  flights$departure <- as.POSIXct(paste(paste(flights$year, flights$month,
                                              flights$day, sep = "-"),
                                        paste(substr(tmp, 1, 2),
                                              substr(tmp, 3, 4),
                                              sep = ":")),
                                  format = "\%Y-\%m-\%d \%H:\%M",
                                  tz = "America/New_York")
  flights$dep_delay <- hms::hms(minutes = flights$dep_delay)
  flights$arr_delay <- hms::hms(minutes = flights$arr_delay)
  flights$air_time <- hms::hms(minutes = flights$air_time)
  flights <- flights[, c("departure", "dep_delay", "arr_delay", "air_time",
                         "carrier", "flight", "tailnum", "origin", "dest",
                         "distance")]
  save(flights, file = "data/flights.rda", compress = "xz")
}

}
\keyword{datasets}
