% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-key.R
\name{ft_key_spec}
\alias{ft_key_spec}
\alias{ft_key_spec.MariaDBConnection}
\title{Generate SQL for fulltext key definition}
\usage{
ft_key_spec(..., con = get_con())

\method{ft_key_spec}{MariaDBConnection}(cols, index_name = NA_character_,
  parser = NA_character_, comment = NA_character_, con = get_con(),
  ...)
}
\arguments{
\item{...}{Arguments passed on to further methods.}

\item{con}{Database connection object.}

\item{cols}{Character vector specifying the column(s) to be used for the
key}

\item{index_name}{(Optional) name of the index.}

\item{parser}{Name of the full-text parser to be used in the index. Examples
are \code{ngram}
(\url{https://dev.mysql.com/doc/refman/5.7/en/fulltext-search-ngram.html})
and \code{mecab}
(\url{https://dev.mysql.com/doc/refman/5.7/en/fulltext-search-mecab.html}).}

\item{comment}{Index definitions can include an optional comment of up to
1024 characters.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for fulltext key definitions
e.g. in CREATE/ALTER TABLE statements.
}
\section{TODO}{
 implement a function show_db_plugins that fetches SHOW
PLUGINS and use the results for checking if the parser is listed as
FTPARSER. Currently the parser is passed without being escaped!
}

