% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cfg}
\alias{get_cfg}
\title{Load config list}
\usage{
get_cfg(file_name = NULL, update = FALSE, section = NULL)
}
\arguments{
\item{file_name}{The file name of the config file. If no value is supplied,
the current wd is searched for files ending in '.yaml' or '.yml', followed
by the extdata dir in the package installation.}

\item{update}{Logical, whether to force reading the config again from file.}

\item{section}{Name of the top-level config file section to return}
}
\value{
The database configuration as a list.
}
