% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spec.R
\name{get_col_spec}
\alias{get_col_spec}
\alias{get_col_spec.integer}
\alias{get_col_spec.integer64}
\alias{get_col_spec.numeric}
\alias{get_col_spec.character}
\alias{get_col_spec.list}
\alias{get_col_spec.blob}
\alias{get_col_spec.logical}
\alias{get_col_spec.factor}
\alias{get_col_spec.POSIXt}
\alias{get_col_spec.Date}
\alias{get_col_spec.difftime}
\alias{get_col_spec.data.frame}
\title{Generate column definition SQL from an R object}
\usage{
get_col_spec(x, ...)

\method{get_col_spec}{integer}(x, ...)

\method{get_col_spec}{integer64}(x, ...)

\method{get_col_spec}{numeric}(x, ...)

\method{get_col_spec}{character}(x, ...)

\method{get_col_spec}{list}(x, ...)

\method{get_col_spec}{blob}(x, ...)

\method{get_col_spec}{logical}(x, ...)

\method{get_col_spec}{factor}(x, ...)

\method{get_col_spec}{POSIXt}(x, ...)

\method{get_col_spec}{Date}(x, ...)

\method{get_col_spec}{difftime}(x, ...)

\method{get_col_spec}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object for which to generate SQL column specification}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Automatically determine the SQL type and create a column
specification from an R object such as a vector or data.frame/tibble.
}
