% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_con}
\alias{get_con}
\title{Get the database connection}
\usage{
get_con(cfg_file = NULL, update = FALSE)
}
\arguments{
\item{update}{Logical, whether to force destroying and recreating the
database connection.}
}
\value{
The database connection object.
}
