% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_ref.R
\name{guess_ref}
\alias{guess_ref}
\title{Guess the format of references/citations from a file}
\usage{
guess_ref(file = file.choose(), return_data = FALSE, quiet = FALSE)
}
\arguments{
\item{file}{(optional) a string indicating a reference/citation file. If not
assigned, a dialog window will be open enabling the user to search and
select a file (only for interactive sessions).}

\item{return_data}{(optional) a \code{logical} value indicating if the function
must return a \code{list} object with the data and metadata from \code{file}. See the
Details section to learn more (default: \code{FALSE}).}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}
}
\value{
\itemize{
\item If \code{return_data = FALSE}, a string with the name of the reference/citation
format found in the \code{file} argument.
\item If \code{return_data = TRUE}, a \code{list} object with the data and metadata from
the \code{file} argument. See the Details section to learn more.
\item The function will return an invisible \code{NULL} if it fails to found a format.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{guess_ref()} try to guess the format of references/citations from a file
(\emph{e.g.}, PubMed, RIS).

At the moment, \code{guess_ref()} works only with PubMed and RIS (Research
Information Systems) formats.
}
\details{
\subsection{\code{return_data} argument}{

When \code{return_data = TRUE}, \code{guess_ref()} will return a \code{list} object with
the following elements:
\itemize{
\item \code{format}: a string with the name of the reference/citation format found in
the \code{file} argument.
\item \code{data}: a \code{character} object representing all the lines found in the \code{file}
argument.
\item \code{count}: A \code{numeric} value with the count of the references/citations found
in the \code{file} argument. If you're only interest in this data, use the
\code{count_ref()} function.
\item \code{index}: an \code{integer} object with values indicating the cut points of the
references/citations. If the cut point is located at the end of the
references/citations, the \code{between} value (see below) will be set to
\code{"right"}, else, it will be set to \verb{"left}.
\item \code{between}: a \code{"left"} or \code{"right"} value that can be used with the
\code{\link[gutils:cutter]{cutter()}} function. See the documentation for the
mentioned function to learn more.
}
}
}
\examples{
\dontrun{
if (require_namespace("utils", quietly = TRUE)) {
    references <- raw_data("reference")
    file <- references[grep("_apa_", references)]
    file <- raw_data("reference", file)
    file <- zip_files <- utils::unzip(file, exdir = tempdir())[1]

    guess_ref(file)
}}
}
\seealso{
Other reference/citation functions: 
\code{\link{count_ref}()},
\code{\link{read_ref}()}
}
\concept{reference/citation functions}
