% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_vec}
\alias{is_vec}
\alias{is_lgl}
\alias{is_num}
\alias{is_int}
\alias{is_chr}
\alias{lt}
\alias{lte}
\alias{gt}
\alias{gte}
\alias{eq}
\title{Test if an object is a vector}
\usage{
is_vec(x, type = c(NA, "int", "num", "lgl", "chr"), n_elem = gte(1L),
  names = FALSE, allow_na = FALSE, allow_null = FALSE)

is_lgl(...)

is_num(...)

is_int(x, ..., strict = TRUE)

is_chr(x, ..., n_char = gte(1L))

lt(b)

lte(b)

gt(b)

gte(b)

eq(b)
}
\arguments{
\item{x}{The object to be tested}

\item{type}{The type of vector that is expected. Possibly choices are
\code{NA}, \code{"int"}, \code{"num"}, \code{"lgl"}, \code{"chr"}, where NA
does not test for a specific type.}

\item{n_elem}{The number of elements that is expected. Using the functions
\code{lt}, \code{lte}, \code{gt}, \code{gte}, \code{eq}, an (in)equality can
be specified where \code{length(x)} is the lhs and the argument passed to
the function, the rhs.}

\item{names}{Either a single logical to test if the object is named or not,
or a character vector used to test for setequal of \code{names(x)} and
\code{names}.}

\item{allow_na}{A logical, specifying whether NA is allowed in x.}

\item{allow_null}{A logical, specifying whether x may be NULL.}

\item{...}{Arguments passed to \code{is_vec}.}

\item{strict}{A logical switch specifying whether a whole number specified
as numeric (instead if integer) will be considered as an integer as well.}

\item{n_char}{Behaves analogously to \code{n_elem}, with \code{nchar(x)} on
the lhs.}

\item{b}{The rhs of the comparison.}
}
\value{
A single logical, indicating whether all conditions are met.
}
\description{
Intended for validating the input to functions, objects can be
tested if they are vectors, satisfying certain additional constraints, such
as type (integer, numeric, etc.), the number of elements, specified as
(in)equality, whether the object is named, whether it contains NA and if it
is NULL.
}
