% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_set.R
\name{keyword_set}
\alias{keyword_set}
\title{Filter a keyword set from the \code{keyword} entity}
\usage{
keyword_set(domain_id, language = NULL, update = FALSE)
}
\arguments{
\item{domain_id}{An \code{integer} or \code{numeric} value indicating the domain id to
return.}

\item{language}{(optional) A string indicating the language constraint of
the keywords (case insensitive).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{keyword_set} returns a tidied single keyword set from the keyword.
}
\examples{
\dontrun{
## __ To write all sheets __
write_sheet()

## __ To write only a specific sheet __
write_sheet(sheet_id()[1])}
}
\seealso{
Other utility functions: 
\code{\link{raw_data}()},
\code{\link{tidy_keyword}()}
}
\concept{utility functions}
