% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_config}
\alias{load_config}
\title{Load yml configuration file}
\usage{
load_config(file_name = NULL)
}
\arguments{
\item{file_name}{The file name of the config file. If no value is supplied,
the current wd is searched for files ending in '.yaml' or '.yml', followed
by the extdata dir in the package installation.}
}
\value{
The database configuration as a list.
}
