% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-compat.R
\name{make_compat}
\alias{make_compat}
\alias{make_compat.MariaDBConnection}
\title{Make column specifications compatible}
\usage{
make_compat(..., con = get_con())

\method{make_compat}{MariaDBConnection}(x, tbl, col = NULL, con = get_con(),
  ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Data to be written to db.}

\item{tbl}{Name of the existing table.}

\item{col}{Column names in the table that are affected.}
}
\value{
Character vector.
}
\description{
For a column to be written to db, test whether the existing
column definition in db is compatible with the new data and if not, create
an appropriate ALTER TABLE clause.
}
