% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_extdata.R
\name{normalize_extdata}
\alias{normalize_extdata}
\title{Normalize file names from the \code{extdata} folder}
\usage{
normalize_extdata(package = NULL)
}
\arguments{
\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{normalize_extdata()} normalize file names founded in the \verb{./inst/extdata/}
folder of a package. \strong{This function must be used only with
\href{https://www.rstudio.com/}{RStudio} IDE}.

The function lowercase the names, substitute \code{" - "} characters for \code{"_"},
and substitute \code{" "} characters for \code{"-"}.
}
\examples{
\dontrun{
normalize_files()}
}
\seealso{
Other SQLR system functions: 
\code{\link{build_reference}()},
\code{\link{keyword_set}()},
\code{\link{read_sheet}()},
\code{\link{start_sqlr}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{SQLR system functions}
