% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-compat.R
\name{parse_col_def}
\alias{parse_col_def}
\alias{parse_col_def.MariaDBConnection}
\title{Unquote identifiers}
\usage{
parse_col_def(..., con = get_con())

\method{parse_col_def}{MariaDBConnection}(x, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Character vector to be parsed.}
}
\value{
Character vector.
}
\description{
Quoted identifiers such as produced by
\code{DBI::dbQuoteIdentifier(con, "foo")} are unquoted and stripped of their
\code{SQL} class, yielding the original character vector (here "foo").
}
\section{TODO}{
 what about character encoding? ie if UTF chars are to be
written to an ascii column: catch that as well?
}

