% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-queries.R
\name{parse_col_spec}
\alias{parse_col_spec}
\alias{parse_col_spec.MariaDBConnection}
\title{Parse SQL column definitions}
\usage{
parse_col_spec(..., con)

\method{parse_col_spec}{MariaDBConnection}(x, ..., con)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Character vector to be parsed.}
}
\value{
A tibble.
}
\description{
Split and parse SQL column definitions, such as created by
[tbl_spec].

\describe{
  \item{Field}{The column name.}
  \item{Type, Length, Unsigned}{See [parse_data_type].}
  \item{NULL}{Logical, whether the column is nullable.}
  \item{Encoding}{The character encoding of a char column.}
  \item{Collation}{The collation of a char column.}
}
}
