% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-queries.R
\name{parse_data_type}
\alias{parse_data_type}
\alias{parse_data_type.MariaDBConnection}
\title{Parse SQL data type definitions}
\usage{
parse_data_type(..., con = get_con())

\method{parse_data_type}{MariaDBConnection}(x, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Character vector to be parsed.}
}
\value{
A tibble.
}
\description{
Split and parse SQL data type definitions, such as created by
any of the data type specification functions (e.g. [col_int], [col_dbl],
etc.). The result is a tibble with as many rows as SQL statements were
passed and with the following columns:

\describe{
  \item{Type}{The column data type.}
  \item{Length}{Either the length (e.g. CHAR(5) yields 5) or the levels
        of an ENUM or SET field.}
  \item{Unsigned}{Logical, whether the numeric column is signed.}
}
}
