% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{parse_sql}
\alias{parse_sql}
\title{Parse SQL}
\usage{
parse_sql(x, con, ...)
}
\arguments{
\item{x}{SQL string}

\item{con}{Database connection}

\item{...}{Generic consistency}
}
\value{
{sqlr} object
}
\description{
The inverse operation as [render_sql()], i.e. instantiating an {sqlr} object
from an SQL string.
}
