% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.R
\docType{data}
\name{planes}
\alias{planes}
\title{Plane metadata.}
\format{A data frame with columns:
\describe{
\item{tailnum}{Tail number}
\item{year}{Year manufactured}
\item{type}{Type of plane}
\item{manufacturer,model}{Manufacturer and model}
\item{engines,seats}{Number of engines and seats}
\item{speed}{Average cruising speed in mph}
\item{engine}{Type of engine}
}}
\source{
https://cran.r-project.org/package=nycflights13, v0.2.2
}
\usage{
planes
}
\description{
Plane metadata for all plane tailnumbers found in the FAA aircraft
registry. American Airways (AA) and Envoy Air (MQ) report fleet numbers
rather than tail numbers (e.g.
\url{http://www.flyerguide.com/Tail_Numbers_(AA)})
so can't be matched.
}
\examples{
\dontrun{
  planes <- nycflights13::planes
  planes$type <- factor(planes$type)
  planes$manufacturer <- factor(planes$manufacturer)
  planes$engine <- factor(planes$engine)
  save(planes, file = "data/planes.rda", compress = "xz")
}

}
\keyword{datasets}
