% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{pubmed}
\alias{pubmed}
\title{Generate a query for PubMed}
\usage{
pubmed(..., constraint = NULL, domain = NULL)
}
\arguments{
\item{...}{One or more \code{character} objects with keywords.}

\item{constraint}{(optional) A \code{character} object indicating the type/types
of constraint of the query.}

\item{domain}{(optional) A \code{list} object with a set of instructions for
keyword conversion.}
}
\value{
A string with a query for PubMed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{pubmed()} returns a query for PubMed.
}
\examples{
\dontrun{
pubmed("Lorem", "Ipsum, dolor", "sit amet", constraint = "abstract")}
}
\seealso{
Other utility functions: 
\code{\link{keyword_set}()},
\code{\link{query}()},
\code{\link{raw_data}()},
\code{\link{tidy_keyword}()}
}
\concept{utility functions}
