% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Generate a query for a database provider}
\usage{
query(provider, domain_set, constraint_set)
}
\arguments{
\item{provider}{A string indicating the database provider name (case
insensitive)}

\item{domain_set}{(optional) A string indicating the domain logic.}

\item{constraint_set}{A string indicating the constraint logic.}
}
\value{
A string with a query for the provider indicating in \code{provider}.
provider.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{query()} returns a query for a specific database provider.
}
\examples{
\dontrun{
query("PubMed", "1 AND 2", "1 AND (4 OR 5 OR 6)")}
}
\seealso{
Other utility functions: 
\code{\link{keyword_set}()},
\code{\link{pubmed}()},
\code{\link{raw_data}()},
\code{\link{tidy_keyword}()}
}
\concept{utility functions}
