% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sheets.R
\name{range_write}
\alias{range_write}
\title{Write a \code{data.frame} object to a review table hosted on Google Sheets}
\usage{
range_write(x, name, package = rutils:::get_package_name(), limit = 2e+05)
}
\arguments{
\item{x}{A \code{data.frame} object with at least 1 row.}

\item{name}{A string indicating the name of the sheets that the
function must write.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}

\item{limit}{(optional) a number with the limit amount of data to write for
each batch. Example: a limit of \code{200000} will write a \code{data.frame} with
\code{10000} rows and \code{50} columns in 3 batches, because \code{10000 * 50 = 500000}
(over the limit) and \code{ceiling(500000 / 200000) = 3} (default: \code{200000}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
\code{sqlr} system.

\code{range_write()} writes a \code{data.frame} object to a specific range of a review
table hosted on Google Sheets.

This function was created as workaround to \link[googlesheets4:range_write]{googlesheets4::range_write}
writing limit. It bypass this limit by writing in batches.

Please note that this function will always write in the \code{A1} range and will
delete all but the 2 first rows of the worksheet before writing on it.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{\link[=write_metadata]{write_metadata()}} to learn more.
}
\examples{
\dontrun{
range_write(reference, "reference")}
}
\seealso{
Other Google Sheets functions: 
\code{\link{read_sheet}()},
\code{\link{sheet_nrow}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{Google Sheets functions}
