% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{sqlr} raw data}
\usage{
raw_data(type = NULL, file = NULL, quiet = FALSE)
}
\arguments{
\item{type}{(optional) a string indicating the file type of the raw dataset
(default: \code{NULL}).}

\item{file}{(optional) a string indicating the file name of the raw dataset.}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}
}
\value{
\itemize{
\item If \code{type = NULL}, a \code{character} object with all file type names
available.
\item If \code{type != NULL} and \code{file = NULL}, a \code{character} object with all file
names available from type.
\item If \code{type != NULL} and \code{file != NULL}, a string file name path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data()} returns the raw data paths of the \code{sqlr} package.
}
\examples{
\dontrun{
raw_data()
}
}
\seealso{
Other utility functions: 
\code{\link{cutter}()},
\code{\link{normalize_extdata}()}
}
\concept{utility functions}
