% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{render_sql}
\alias{render_sql}
\title{Generate SQL}
\usage{
render_sql(x, con, ...)
}
\arguments{
\item{x}{Object}

\item{con}{Database connection}
}
\value{
SQL string
}
\description{
For a given SQL dialect determined by `con`, SQL is generated corresponding
to the object passed as `x`.
}
