% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ris_tags}
\alias{ris_tags}
\title{List of search tags from several database providers}
\format{
An object of class \code{list} of length 4.
}
\usage{
ris_tags
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{UNDER DEVELOPMENT}
}
\keyword{datasets}
