% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_con}
\alias{rm_con}
\title{Destroy the database connection}
\usage{
rm_con()
}
\value{
NULL (invisibly)
}
\description{
Destroy the database connection object that is currently saved
in the config environment. It will only be destroyed on the next garbage
collection run.
}
