\name{set}
\alias{set}
\alias{set.character}
\alias{set.data.frame}
\title{SQL Set Statement}
\usage{
  set(x, ...)

  \method{set}{character} (...)

  \method{set}{data.frame} (x)
}
\arguments{
  \item{x}{}
}
\description{
  Update table and set columns to specific values.
}
\author{
  Ciaran Tobin
}

