% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-con.R
\name{set_con}
\alias{set_con}
\title{Set the database connection}
\usage{
set_con(file_name = NULL, section = "db_setup", update = FALSE)
}
\arguments{
\item{file_name}{The file name of the config file. If no value is supplied,
the current wd is searched for files ending in '.yaml' or '.yml', followed
by the extdata dir in the package installation.}

\item{section}{A list node to be returned instead of the whole config list.
If multiple nodes match, the last match (the list is traversed recursively)
will be returned}

\item{update}{Logical switch for forcing a reset to the con object despite
it being present and valid.}
}
\value{
TRUE if the con object was set, FALSE if not.
}
\description{
Using the information from the yml config file, generate a
data base connection object and store it alongside the config information
in the config environment.
}
