% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sheets.R
\name{sheet_id}
\alias{sheet_id}
\title{Get the \code{sheet_id} of database entities on Google Sheets}
\usage{
sheet_id(name = NULL, package = NULL)
}
\arguments{
\item{name}{(optional) A string indicating the sheet name (default: \code{NULL}).}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}
}
\value{
If \code{name = NULL}, returns a \code{character} object with all sheet names
available. Else, it returns the \code{sheet_id} object of the sheet indicated in
\code{name}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{sheet_id()} retrieve the \code{sheet_id} from entities tables on Google Sheets.
}
\examples{
\dontrun{
## __ To list all the all sheet names available __
sheet_id()

## __ To get the 'sheet_id' object of a specific sheet __
sheet_id(sheet_id()[1])}
}
\seealso{
Other GIPSO functions: 
\code{\link{keyword_set}()},
\code{\link{read_sheet}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{GIPSO functions}
