% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-queries.R
\name{show_db_cols}
\alias{show_db_cols}
\title{Generate SQL for a boolean data type definition}
\usage{
show_db_cols(..., con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for boolean data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
