% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{data}
\name{source}
\alias{source}
\title{Source dataset}
\format{
An object of class \code{function} of length 1.
}
\usage{
source
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{UNDER DEVELOPMENT}
}
\keyword{datasets}
