% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-key.R
\name{spat_key_spec}
\alias{spat_key_spec}
\title{Generate SQL for spatial key definition}
\usage{
spat_key_spec(cols, index_name = NA_character_,
  comment = NA_character_, ...)
}
\arguments{
\item{cols}{Character vector specifying the column(s) to be used for the
key}

\item{index_name}{(Optional) name of the index.}

\item{comment}{Index definitions can include an optional comment of up to
1024 characters.}

\item{...}{Arguments passed on to further methods.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for spatial key definitions
e.g. in CREATE/ALTER TABLE statements.
}
