% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-tbl.R
\name{spec_lst}
\alias{spec_lst}
\title{Wrap multiple specs into a list}
\usage{
spec_lst(..., con = get_con())
}
\arguments{
\item{...}{Function calls to functions that create spec objects, such as
\code{col_spec} or \code{key_spec}. All must evaluate to SQL strings.}

\item{con}{A connection used to determine the SQL dialect to be used}
}
\value{
A list of SQL strings.
}
\description{
Instead of using \code{list} to combine multiple spec objects
such as produced by \code{col_spec}, this function passes its connection
object to the enclosed function calls.
}
