% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssql.R
\name{sql_conn_common}
\alias{sql_conn_common}
\title{设置通用的数据库连接}
\usage{
sql_conn_common(
  ip = "115.159.201.178",
  port = 1433,
  user_name = "sa",
  password = "Hoolilay889@",
  db_name = "JH_2018B"
)
}
\arguments{
\item{ip}{服务器地址}

\item{port}{服务器端口}

\item{user_name}{用户名}

\item{password}{密码}

\item{db_name}{数据库名称}
}
\value{
返回值
}
\description{
设置通用的数据库连接
}
\examples{
sql_conn_common()
}
