% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssql.R
\name{sql_tableCount}
\alias{sql_tableCount}
\title{返回表的行数}
\usage{
sql_tableCount(conn, table_name)
}
\arguments{
\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
返回表的行数
}
\examples{
hana_tableCount()
}
