% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssql.R
\name{sql_update}
\alias{sql_update}
\title{更新update数据}
\usage{
sql_update(conn, sql_str)
}
\arguments{
\item{conn}{连接信息}

\item{sql_str}{sql信息}
}
\value{
返回值
}
\description{
更新update数据
}
\examples{
sql_update();
}
