% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssql.R
\name{sql_writeTable}
\alias{sql_writeTable}
\title{将R对象写入到SQL_server表中}
\usage{
sql_writeTable(conn, table_name, r_object, append = FALSE)
}
\arguments{
\item{conn}{连接信息}

\item{table_name}{表名}

\item{r_object}{R对象表}

\item{skip}{是否跳过表名重复检查，默认为是，不需要重复性}
}
\value{
返回T表示插入成功，F可能是表名重复等
}
\description{
将R对象写入到SQL_server表中
}
\examples{
db_saveR2Table();
}
