% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr-package.R
\docType{package}
\name{sqlr-package}
\alias{sqlr}
\alias{sqlr-package}
\title{sqlr: Represent and Generate SQL Schemas}
\description{
A dialect-independent representation of a relational schema -- tables, columns, types, constraints and indexes -- that can be built up from R, rendered as data definition language for a given SQL dialect, and learned back from a live database. Dialect support is provided by companion packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nbenn/sqlr}
  \item Report bugs at \url{https://github.com/nbenn/sqlr/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{r@nbenn.ch}

Authors:
\itemize{
  \item Nicolas Bennett \email{r@nbenn.ch}
}

}
\keyword{internal}
