% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sqlr_quote}
\alias{sqlr_quote}
\alias{sqlr_quote_literal}
\title{Quote identifiers and literals}
\usage{
sqlr_quote(dialect, ...)

sqlr_quote_literal(dialect, ...)
}
\arguments{
\item{dialect}{A [sqlr_dialect].}

\item{...}{The identifier or literal to quote, passed to methods.}
}
\value{
A string.
}
\description{
Quote identifiers and literals
}
