% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sql.R
\name{sqlr_sql}
\alias{sqlr_sql}
\title{Verbatim SQL}
\usage{
sqlr_sql(text = character(0))
}
\arguments{
\item{text}{String of SQL.}
}
\value{
An object of class `sqlr_sql`.
}
\description{
Wraps a string that is to be emitted as-is, without quoting or escaping.
Accepted anywhere a value, expression or type may appear.
}
\examples{
sqlr_sql("now()")

}
