% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_sqlr.R
\name{start_sqlr}
\alias{start_sqlr}
\title{Start settings for SQLR R package templates}
\usage{
start_sqlr(id, sheet = "Dataset", package = NULL)
}
\arguments{
\item{id}{A string with the Google Sheets ID from the 'Sheets' table.}

\item{sheet}{(optional) a string indicating the worksheet/tab where the
sheets data can be found on the 'Sheets spreadsheet (default: \code{"Dataset"}).}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{start_sqlr()} handles the initial commands to start Systematic Quantitative
Literature Reviws R packages.
}
\seealso{
Other data functions: 
\code{\link{write_metadata}()}
}
\concept{data functions}
