% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-spec.R
\name{tbl_spec}
\alias{tbl_spec}
\alias{tbl_spec.MariaDBConnection}
\title{Generate SQL for a table definition}
\usage{
tbl_spec(..., con = get_con())

\method{tbl_spec}{MariaDBConnection}(name = paste(sample(letters, 10, TRUE),
  collapse = ""), temp = FALSE, force = FALSE,
  cols = spec_lst(col_spec(name = "id", type = "int", nullable = FALSE,
  auto_increment = TRUE, key = "primary", unsigned = TRUE)), keys = NULL,
  engine = "InnoDB", auto_incr = NA_integer_, char_set = NA_character_,
  collate = NA_character_, comment = NA_character_, table_options = NULL,
  partition = NULL, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{name}{String specifying the table name}

\item{temp}{Logical switch specifying whether the table is temporary
(visible only within the current session, and dropped automatically when
the session is closed)}

\item{force}{Logical, if FALSE, an error is thrown if the table exists, if
TRUE, nothing happens if the table already exists.}

\item{cols, keys}{Column and key specifications as generated by
\code{col_spec} and the various key spec functions. If multiple
columns/keys are passed, wrap function calls in \code{spec_lst}.}

\item{engine}{The storage engine (defaults to InnoDB)}

\item{auto_incr}{The initial \code{AUTO_INCREMENT} value for the table.}

\item{char_set}{Specifies a default character set for the table.}

\item{collate}{Specifies a default collation for the table.}

\item{comment}{A comment for the table, up to 2048 characters long.}

\item{table_options}{Further table options such as \code{COMPRESSION},
\code{ENCRYPTION}, \code{ROW_FORMAT}, etc. Must be one or more SQL strings
and multiple will end up separated by single spaces.}

\item{partition}{Table partition settings. Currently only NULL is accepted.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for table definitions in 
CREATE TABLE statements.
}
\section{TODO}{
 create single \code{key_spec} function like \code{col_spec}

 implement table partitioning
}

