% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_keyword.R
\name{tidy_keyword}
\alias{tidy_keyword}
\title{Tidy keywords}
\usage{
tidy_keyword(
  ...,
  min_chars = 1,
  delimiter = ",",
  clean_modifiers = TRUE,
  sort = FALSE,
  na_rm = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{...}{One or more \code{character} objects with keywords.}

\item{min_chars}{(optional) A number indicating the minimal number of
characters a keyword must have. Keywords that don't comply to this setting
will be transformed to \code{NA}.}

\item{delimiter}{(optional) A string of length 1 with the delimiter used on
\code{...} values. Use \code{delimiter = NULL} to disable this behavior (default:
\code{NULL}).}

\item{clean_modifiers}{(optional) A \code{logical} value indicating if keywords
with modifiers must be transformed to \code{NA}. Modifiers follow the
\verb{[*?():'\\"]} pattern (default: \code{TRUE}).}

\item{sort}{(optional) A \code{logical} value indicating if the output must be
sorted (default: \code{FALSE}).}

\item{na_rm}{(optional) A \code{logical} value indicating if \code{NA} values must be
removed from the output (default: \code{TRUE}).}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}
}
\value{
A \code{character} object with keywords in \code{...} tidied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{tidy_keyword()} tidy keyword values for database search.
}
\details{
\code{tidy_keyword()} will automatically lowercase all objects in \code{...}.
}
\examples{
tidy_keyword("Lorem", "Ipsum dolor", c("sit", "amet"))
#> [1] "lorem"           "\"ipsum dolor\"" "sit" # Expected
#> [4] "amet" # Expected

tidy_keyword("Lorem", "Ipsum dolor", c("sit", "amet"), sort = TRUE)
#> [1] "amet"            "\"ipsum dolor\"" "lorem" # Expected
#> [4] "sit" # Expected

tidy_keyword("Lorem", "Ipsum, dolor", delimiter = ",")
#> [1] "lorem" "ipsum" "dolor" # Expected

tidy_keyword("Lorem", "Ipsum*", clean_modifiers = TRUE, na_rm = TRUE)
#> [1] "lorem" # Expected

tidy_keyword("Lorem", "Ipsum*", clean_modifiers = TRUE, na_rm = FALSE)
#> [1] "lorem" NA # Expected
}
\seealso{
Other utility functions: 
\code{\link{keyword_set}()},
\code{\link{pubmed}()},
\code{\link{query}()},
\code{\link{raw_data}()}
}
\concept{utility functions}
