% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-compat.R
\name{unquote_ident}
\alias{unquote_ident}
\alias{unquote_ident.MariaDBConnection}
\title{Unquote identifiers}
\usage{
unquote_ident(..., con = get_con())

\method{unquote_ident}{MariaDBConnection}(x, con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Character vector to be unquoted.}
}
\value{
Character vector.
}
\description{
Quoted identifiers such as produced by
\code{DBI::dbQuoteIdentifier(con, "foo")} are unquoted and stripped of their
\code{SQL} class, yielding the original character vector (here "foo").
}
