% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-queries.R
\name{unquote_str}
\alias{unquote_str}
\alias{unquote_str.MariaDBConnection}
\title{Unquote strings}
\usage{
unquote_str(..., con = get_con())

\method{unquote_str}{MariaDBConnection}(x, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{x}{Character vector to be unquoted.}
}
\value{
Character vector.
}
\description{
Quoted strings such as produced by
\code{DBI::dbQuoteString(con, "foo")} are unquoted and stripped of their
\code{SQL} class, yielding the original character vector (here "foo").
}
