% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_trial.R
\name{update_reference}
\alias{update_reference}
\title{Update the \code{reference} table using a trial table}
\usage{
update_reference(trial_id, package = rutils:::get_package_name(), write = TRUE)
}
\arguments{
\item{trial_id}{A string indicating the ID of the trial to be written.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}

\item{write}{(optional) a \code{logical} value indicating if the function must
update the \code{reference} data table of the package and to write the new
reference table to the reference spreadsheet listed on the \code{sqlr::sheets}
object (default: \code{TRUE}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{update_reference()} updates the \code{criteria_id}, \code{trial_id}, and \code{pdf} columns
of the \code{reference} table using the data available in a trial table.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{\link[=write_metadata]{write_metadata()}} to learn more.
}
\examples{
\dontrun{
update_reference("NR1")}
}
\seealso{
Other reference/citation functions: 
\code{\link{build_reference}()},
\code{\link{write_reference}()},
\code{\link{write_trial}()}
}
\concept{reference/citation functions}
