% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.R
\docType{data}
\name{weather}
\alias{weather}
\title{Hourly weather data}
\format{A data frame with columns
\describe{
\item{origin}{Weather station. Named origin to faciliate merging with
  \code{\link{flights}} data}
\item{ts}{Date/time of recording as a \code{POSIXct} date}
\item{temp,dewp}{Temperature and dewpoint in F}
\item{humid}{Relative humidity}
\item{wind_dir,wind_speed,wind_gust}{Wind direction (in degrees), speed
  and gust speed (in mph)}
\item{precip}{Preciptation, in inches}
\item{pressure}{Sea level pressure in millibars}
\item{visib}{Visibility rating, corresponds to rounded up miles (0-10)}
}}
\source{
ASOS download from Iowa Environmental Mesonet,
  https://mesonet.agron.iastate.edu/request/download.phtml.
}
\usage{
weather
}
\description{
Hourly meterological data for LGA, JFK and EWR.
}
\examples{
\dontrun{
  weather <- nycflights13::weather
  weather <- weather[weather$month \%in\% 1:3, ]
  weather$ts <- as.POSIXct(paste(paste(weather$year, weather$month,
                                       weather$day, sep = "-"),
                                 paste0(sprintf("\%02d", weather$hour),
                                        ":00")),
                                 format = "\%Y-\%m-\%d \%H:\%M",
                                 tz = "America/New_York")
  weather <- weather[!is.na(weather$ts), ]
  weather[, c("temp", "dewp", "humid", "precip", "pressure")] <- lapply(
    weather[, c("temp", "dewp", "humid", "precip", "pressure")], as.single)
  weather$wind_dir <- as.integer(weather$wind_dir)
  weather$visib <- factor(ceiling(weather$visib), levels = 0:10,
                          ordered = TRUE)
  weather <- weather[, c("origin", "ts", "temp", "dewp", "humid",
                         "wind_dir", "wind_gust", "wind_speed", "precip",
                         "pressure", "visib")]
  save(weather, file = "data/weather.rda", compress = "xz")
}

}
\keyword{datasets}
