% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_db_tbl}
\alias{write_db_tbl}
\alias{write_db_tbl.MariaDBConnection}
\title{Add a data.frame to database}
\usage{
write_db_tbl(..., con = get_con())

\method{write_db_tbl}{MariaDBConnection}(name, data, mode = c("insert",
  "replace"), priority = c(NA, "low", "high", "delayed"), ignore = FALSE,
  partition = NULL, on_dupl_key = NULL, ..., con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods, some of which might end up
as arguments to [create_db_tbl].}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{name}{String specifying the table name}

\item{data}{A data.frame, holding the data to be added. Columns have to be
named.}

\item{mode}{Either \code{"insert"} or \code{"replace"}, specifying the
behavior in case of unique key constraint failures.}

\item{priority}{One of \code{NA}, \code{"low"}, \code{"high"},
\code{"delayed"}. In low priority mode, execution of the \code{INSERT} is
delayed until no other clients are reading from the table. High priority
mode overrides the effect of the \code{--low-priority-updates} option if the
server was started with that option. Both only affect storage engines that
use only table-level locking (e.g. MyISAM). Delayed inserts are deprecated.}

\item{ignore}{Logical switch, specifying whether errors that occur while
executing the \code{INSERT} statement are ignored. For example a unique key
violation will instead of throwing an error, only throw a warning and
discard the offending row to be inserted.}

\item{partition}{Intended for controlling the behavior of inserts into
partitioned tabled, this is currently not implemented.}

\item{on_dupl_key}{If not \code{NULL} and a row is inserted that causes a
cause a unique key constraint failure, an UPDATE of the old row occurs. A
named list is expected where names correspond to columns to be updated and
list entries are expected to be SQL expressions and of type \code{SQL}.}
}
\value{
The number of added rows (invisibly).
}
\description{
Add data contained in a \code{data.frame} to database. In case
the table does not exist, it will be created by [create_db_tbl], using
the arguments passed as \code{...}.
}
