% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sheets.R
\name{write_sheet}
\alias{write_sheet}
\title{Write sheets of database entities on Google Sheets to the package}
\usage{
write_sheet(name = NULL, package = NULL)
}
\arguments{
\item{name}{(optional) A string indicating the sheet name (default: \code{NULL}).}

\item{package}{(optional) A string indicating the package where the sheets
list can be found. If \code{NULL}, the function will try to use the name of the
project active directory (requires the \code{rstudioapi} package) (default:
\code{NULL}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{write_sheet} reads entities tables on Google Sheets and write them in the
data directory of an R package.
}
\examples{
\dontrun{
## __ To write all sheets __
write_sheet()

## __ To write only a specific sheet __
write_sheet(sheet_id()[1])}
}
\seealso{
Other GIPSO functions: 
\code{\link{keyword_set}()},
\code{\link{read_sheet}()},
\code{\link{sheet_id}()}
}
\concept{GIPSO functions}
