% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference.R
\name{build_reference}
\alias{build_reference}
\title{Build and write the \code{reference} table to a SQLR package}
\usage{
build_reference(package = rutils:::get_package_name(), write = TRUE)
}
\arguments{
\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}

\item{write}{(optional) a \code{logical} value indicating if the function must
write a \code{reference.rda} file to \code{"./data/"} and also write to the reference
spreadsheet listed on the \code{sqlr::sheets} object (default: \code{TRUE}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{build_reference()} builds and write the \code{reference} entity table of
the Systematic Quantitative Literature Review (SQLR) system to an R package.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{\link[=write_metadata]{write_metadata()}} to learn more.
}
\examples{
\dontrun{
build_reference()}
}
\seealso{
Other reference/citation functions: 
\code{\link{update_reference}()},
\code{\link{write_reference}()},
\code{\link{write_trial}()}
}
\concept{reference/citation functions}
