% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_chr}
\alias{col_chr}
\title{Generate SQL for a string data type definition}
\usage{
col_chr(length = 255L, fixed = FALSE, force_text = length >= 16384L,
  char_set = NA_character_, collate = NA_character_, ...)
}
\arguments{
\item{length}{The maximum length of string the column is expected to hold.}

\item{fixed}{Logical switch specifying whether strings are fixed length
or varying in length.}

\item{force_text}{String length threshold for forcing the type from
(VAR)CHAR to TEXT which causes the values to be stored separately from the
rest of the row. TEXT columns only contribute 9 to 12 bytes toward the row
size limit (65,535 bytes).}

\item{char_set, collate}{The character set and collation used.}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for string data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
\section{TODO}{
 implement a function show_db_charset that fetches SHOW
CHARACTER SET and use the results for checking if the char_set is listed.
Furthermore, SHOW COLLATION WHERE Charset = 'char_set' can be used to test
whether the selected collation is allowed for the given charset. A warning
could be generated if the user selects a charset with Maxlen > 1: max row
size becomes smaller and char wastes space
}

