% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-spec.R
\name{col_dbl}
\alias{col_dbl}
\title{Generate SQL for floating point data type definition}
\usage{
col_dbl(..., con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}
}
\value{
SQL to be used in a CREATE table statement
}
