% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_dbl}
\alias{col_dbl}
\title{Generate SQL for floating point data type definition}
\usage{
col_dbl(prec = "double", unsigned = FALSE, ...)
}
\arguments{
\item{prec}{One of \code{single}, \code{double}, specifying the floating
point precision the column is expected to be capable of holding.}

\item{unsigned}{Logical switch specifying whether the values are singed or
unsigned.}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for floating point data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
