% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_dtm}
\alias{col_dtm}
\title{Generate SQL for a date/time data type definition}
\usage{
col_dtm(class = c("datetime", "date", "time", "year"), val = NULL, ...)
}
\arguments{
\item{class}{One of \code{datetime}, \code{date}, \code{time}, \code{year},
specifying the type of date/time of the column.}

\item{val}{A value, the class of which is used for automatically inferring
type.}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for date/time data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
