% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-spec.R
\name{col_dtm}
\alias{col_dtm}
\title{Generate SQL for a date/time data type definition}
\usage{
col_dtm(..., con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for date/time data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
