% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_fct}
\alias{col_fct}
\alias{col_fct.MariaDBConnection}
\title{Generate SQL for a factor data type definition}
\usage{
col_fct(..., con = get_con())

\method{col_fct}{MariaDBConnection}(levels, variant = c("enum", "set"),
  char_set = NA_character_, collate = NA_character_, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{levels}{A character vector, specifying the factor levels.}

\item{variant}{One of \code{enum}, \code{set}.}

\item{char_set, collate}{The character set and collation used.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for factor data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
