% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-col.R
\name{col_int}
\alias{col_int}
\title{Generate SQL for an integer data type definition}
\usage{
col_int(size = "int", unsigned = FALSE, min = NA, max = NA, ...)
}
\arguments{
\item{size}{One of \code{tiny}, \code{small}, \code{medium}, \code{int},
\code{big}, specifying the size of integer.}

\item{unsigned}{Logical switch specifying whether the integer is singed or
unsigned.}

\item{min, max}{The minimum/maximum values the integer column has to be able
to hold. This can be used for inferring size and signedness automatically.}

\item{...}{Arguments passed to the S3 methods}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for integer data type
specifications in column definitions for CREATE/ALTER TABLE statements.
}
