% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-spec.R
\name{col_spec}
\alias{col_spec}
\title{Generate SQL for a column definition}
\usage{
col_spec(..., con = get_con())
}
\arguments{
\item{...}{Arguments passed to the S3 methods}

\item{con}{A connection used to determine the SQL dialect to be used}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for column definitions in 
CREATE/ALTER TABLE statements.
}
