% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_mysql}
\alias{connect_mysql}
\title{Connect to a MySQL database}
\usage{
connect_mysql(db_name, username, password, host = "localhost", port = 3306,
  ...)
}
\arguments{
\item{db_name}{The name of the database to connect to/create}

\item{password}{The password used to authenticate the user with the
database}

\item{host}{Hostname of the MySQL server (default: "localhost")}

\item{port}{The port of the MySQL server (default: 3306)}

\item{user}{The username with which to connect to the database}
}
\value{
Connection object that can be used for further database access.
}
\description{
This function provides a wrapper around \code{dbConnect} that
sets up a database and user with all privileges to that schema if any of the
two does not already exist.
}
