% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-con.R
\name{get_con}
\alias{get_con}
\title{Get the database connection}
\usage{
get_con(...)
}
\arguments{
\item{...}{Arguments passed to [set_con].}
}
\value{
The database connection object.
}
\description{
Fetch the database connection stored in the config environment.
If no valid con object is present, it will be created using [set_con()].
}
