% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_set.R
\name{keyword_set}
\alias{keyword_set}
\title{Get a keyword set from the \code{keyword} dataset}
\usage{
keyword_set(domain_id, language = NULL, update = FALSE)
}
\arguments{
\item{domain_id}{An \code{integer} or \code{numeric} value indicating the domain ID to
return.}

\item{language}{(optional) A string indicating the language constraint of
the keywords (case insensitive).}

\item{update}{A \code{logical} value indicating if the \code{keyword} dataset must
be updated (default: \code{FALSE}).}
}
\value{
A \code{character} object with a tidied single keyword set from the
keyword dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{keyword_set} returns a tidied single keyword set from the keyword.
}
\examples{
\dontrun{
keyword_set(1, "english")}
}
\seealso{
Other utility functions: 
\code{\link{pubmed}()},
\code{\link{query}()},
\code{\link{raw_data}()},
\code{\link{tidy_keyword}()}
}
\concept{utility functions}
