% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_set.R
\name{keyword_set}
\alias{keyword_set}
\title{Group keywords from the \code{keyword} dataset}
\usage{
keyword_set(domain_id, language = NULL, package = NULL)
}
\arguments{
\item{domain_id}{An \code{integer} or \code{numeric} value indicating the domain ID to
return.}

\item{language}{(optional) A string indicating the language constraint of
the keywords (case insensitive).}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}
}
\value{
A \code{character} object with a keyword set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{keyword_set()} returns a keyword set from the \code{keyword} dataset.
}
\examples{
\dontrun{
keyword_set(1, "english")}
}
\seealso{
Other GIPSO functions: 
\code{\link{read_sheet}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{GIPSO functions}
