% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_extdata.R
\name{normalize_extdata}
\alias{normalize_extdata}
\title{Normalize file names from the \code{extdata} folder}
\usage{
normalize_extdata()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{normalize_extdata()} normalize file names founded in the \verb{./inst/extdata/}
folder of a package. \strong{This function must be used only with
\href{https://www.rstudio.com/}{RStudio} IDE}.

The function lowercase the names, substitute \code{" - "} characters for \code{"_"},
and substitute \code{" "} characters for \code{"-"}.
}
\examples{
\dontrun{
normalize_files()}
}
\seealso{
Other utility functions: 
\code{\link{cutter}()},
\code{\link{raw_data}()}
}
\concept{utility functions}
