% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{sqlr.pregnancy} raw datasets}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{A string indicating the file name of the raw dataset. If \code{NULL},
all raw dataset file names will be printed (default: \code{NULL}).}
}
\value{
If \code{path = NULL}, returns a \code{character} object with all raw dataset
file names available. Else, it returns the \code{file} path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data()} returns the raw data paths from the \code{sqlr.pregnancy} package.
}
\examples{
\dontrun{
## __ To list all raw dataset file names available __
raw_data()

## __ To get the file path from a specific raw dataset __
raw_data(raw_data()[1])}
}
\seealso{
Other utility functions: 
\code{\link{keyword_set}()},
\code{\link{tidy_keyword}()}
}
\concept{utility functions}
