% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sheets.R
\name{sheet_nrow}
\alias{sheet_nrow}
\title{Get the number of rows from a review table hosted on Google Sheets}
\usage{
sheet_nrow(name, package = gutils:::get_package_name(), rm_header = TRUE)
}
\arguments{
\item{name}{A string indicating the name of the sheets that the
function must evaluate.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}

\item{rm_header}{(optional) A \code{logical} value indicating if the function
must consider the header when computing the number of rows. This only works
when the sheets have one line as header (default: \code{TRUE}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{sheet_nrow()} returns the number of rows of a review table hosted on Google
Sheets.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{\link[=write_metadata]{write_metadata()}} to learn more.
}
\examples{
\dontrun{
sheet_nrow(sheets$domain$name)}
}
\seealso{
Other Google Sheets functions: 
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()}
}
\concept{Google Sheets functions}
