% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-queries.R
\name{show_db_cols}
\alias{show_db_cols}
\alias{show_db_cols.MariaDBConnection}
\title{Show columns of db table(s)}
\usage{
show_db_cols(..., con = get_con())

\method{show_db_cols}{MariaDBConnection}(tbl, like = NULL, where = NULL,
  full = FALSE, parse = FALSE, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed on to further methods.}

\item{con}{A connection object to connect to the db.}

\item{tbl}{The table name.}

\item{like}{A string, specifying a pattern to match column names against.}

\item{where}{An SQL expression used to filter results.}

\item{full}{A logical switch used to request additional information on each
of the returned columns.}

\item{parse}{Logical switch for turning the results into a SequelPro
inspired "Structure" view.}
}
\value{
A tibble.
}
\description{
Display information about the columns in a given table.
The result may be limited using the arguments \code{like}/\code{where}
and more complete information can be requested using the switch \code{full}.
The returned tibble holds the following values for each table column (see
\url{https://dev.mysql.com/doc/refman/5.7/en/show-columns.html}):

\describe{
  \item{Field}{The column name.}
  \item{Type}{The column data type.}
  \item{Collation}{The collation for nonbinary string columns, or NULL for
        other columns. This value is displayed only if you use the FULL
        keyword.}
  \item{Null}{Column nullability. The value is YES if NULL values can be
        stored in the column, NO if not.}
  \item{Key}{Whether the column is indexed: one of NA, PRI, UNI, MUL.}
  \item{Default}{The default value for the column. This is NULL if the
        column has an explicit default of NULL, or if the column definition
        includes no DEFAULT clause.}
  \item{Extra}{Any additional information that is available about a given
        column, such as \code{auto_increment}}
  \item{Privileges}{The privileges you have for the column. This value is
        displayed only if \code{full} is true.}
  \item{Comment}{Any comment included in the column definition. This value
        is displayed only if \code{full} is true.}
}
}
