% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-spec.R
\name{spat_key_spec}
\alias{spat_key_spec}
\alias{spat_key_spec.MariaDBConnection}
\title{Generate SQL for spatial key definition}
\usage{
spat_key_spec(..., con = get_con())

\method{spat_key_spec}{MariaDBConnection}(cols, index_name = NA_character_,
  comment = NA_character_, con = get_con(), ...)
}
\arguments{
\item{...}{Arguments passed on to further methods.}

\item{con}{Database connection object.}

\item{cols}{Character vector specifying the column(s) to be used for the
key}

\item{index_name}{(Optional) name of the index.}

\item{comment}{Index definitions can include an optional comment of up to
1024 characters.}
}
\value{
SQL to be used in a CREATE table statement
}
\description{
Generate SQL, that can be used for spatial key definitions
e.g. in CREATE/ALTER TABLE statements.
}
