% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr-package.R
\docType{package}
\name{sqlr-package}
\alias{sqlr}
\alias{sqlr-package}
\title{sqlr: A System for Systematic Quantitative Literature Reviews}
\description{
A system for systematic quantitative literature reviews.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/sqlr/}
  \item \url{https://github.com/giperbio/sqlr/}
  \item Report bugs at \url{https://github.com/giperbio/sqlr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
