% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr-package.R
\docType{package}
\name{sqlr-package}
\alias{sqlr}
\alias{sqlr-package}
\title{sqlr: Build SQL queries for creating/altering tables}
\description{
Starting with MardaDB/MySQL and possibly expanding to other SQL dialects, the aim of this package is to help with generating queries for creating and altering tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nbenn/sqlr}
  \item Report bugs at \url{https://github.com/nbenn/sqlr/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{r@nbenn.ch}

}
\keyword{internal}
