% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R, R/spec-col.R, R/spec-key.R, R/spec-tbl.R
\name{sqlr_render}
\alias{sqlr_render}
\alias{sqlr_render.sqlr_col_spec}
\alias{sqlr_render.sqlr_col_int}
\alias{sqlr_render.sqlr_col_dbl}
\alias{sqlr_render.sqlr_col_chr}
\alias{sqlr_render.sqlr_col_raw}
\alias{sqlr_render.sqlr_col_lgl}
\alias{sqlr_render.sqlr_col_fct}
\alias{sqlr_render.sqlr_col_dtm}
\alias{sqlr_render.sqlr_fk_spec}
\alias{sqlr_render.sqlr_pk_spec}
\alias{sqlr_render.sqlr_uk_spec}
\alias{sqlr_render.sqlr_key_spec}
\alias{sqlr_render.sqlr_ft_key_spec}
\alias{sqlr_render.sqlr_spat_key_spec}
\alias{sqlr_render.sqlr_tbl_spec}
\title{Create an SQL representation of an object}
\usage{
sqlr_render(x, con, ...)

\method{sqlr_render}{sqlr_col_spec}(x, con, ...)

\method{sqlr_render}{sqlr_col_int}(x, con, ...)

\method{sqlr_render}{sqlr_col_dbl}(x, con, ...)

\method{sqlr_render}{sqlr_col_chr}(x, con, ...)

\method{sqlr_render}{sqlr_col_raw}(x, con, ...)

\method{sqlr_render}{sqlr_col_lgl}(x, con, ...)

\method{sqlr_render}{sqlr_col_fct}(x, con, ...)

\method{sqlr_render}{sqlr_col_dtm}(x, con, ...)

\method{sqlr_render}{sqlr_fk_spec}(x, con, ...)

\method{sqlr_render}{sqlr_pk_spec}(x, con, ...)

\method{sqlr_render}{sqlr_uk_spec}(x, con, ...)

\method{sqlr_render}{sqlr_key_spec}(x, con, ...)

\method{sqlr_render}{sqlr_ft_key_spec}(x, con, ...)

\method{sqlr_render}{sqlr_spat_key_spec}(x, con, ...)

\method{sqlr_render}{sqlr_tbl_spec}(x, con, ...)
}
\arguments{
\item{x}{An object of class `sqlr`}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{...}{For extensibility.}
}
\description{
TBD.
}
