% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{sqlr_render}
\alias{sqlr_render}
\title{Create an SQL representation of an object}
\usage{
sqlr_render(x, con, ...)
}
\arguments{
\item{x}{An object of class `sqlr`}

\item{con}{A connection used to determine the SQL dialect to be used}

\item{...}{For extensibility.}
}
\description{
TBD.
}
