% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sheets.R
\name{write_metadata}
\alias{write_metadata}
\title{Create a list with Google Sheets metadata for tables hosted on the platform}
\usage{
write_metadata(id, sheet = "Dataset")
}
\arguments{
\item{id}{A string with the Google Sheets ID from the \code{Sheets} table.}

\item{sheet}{(optional) a string indicating the worksheet/tab where the
sheets data can be found in the \code{Sheets} spreadsheet (default:
\code{"Dataset"}).}
}
\value{
An invisible \code{list} object containing lists with the Google Sheets
metadata of the review tables hosted on the platform.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{write_metadata()} creates and returns a \code{list} object containing lists with
the Google Sheets metadata of the review tables hosted on the platform.

This must be set for all R packages from GIPSO for systematic quantitative
literature reviews.

You can see a example of a metadata sheets in: \url{https://bit.ly/2PFWhev}.
}
\examples{
\dontrun{
write_metadata("1x6Aj8cXl9qFtpXq48Q6zdmT-w9EEqNG1iVQcKVRhgKM")
}
}
\seealso{
Other SQLR system functions: 
\code{\link{build_reference}()},
\code{\link{keyword_set}()},
\code{\link{normalize_extdata}()},
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{sheet_nrow}()},
\code{\link{start_sqlr}()},
\code{\link{update_reference}()},
\code{\link{write_query}()},
\code{\link{write_reference}()},
\code{\link{write_sheet}()},
\code{\link{write_trial}()}
}
\concept{SQLR system functions}
