% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_query.R
\name{write_query}
\alias{write_query}
\title{Write queries in the \code{search} dataset}
\usage{
write_query(range = NULL, package = gutils:::get_package_name())
}
\arguments{
\item{range}{(optional) a string indicating the cell range or start of the
target rectangle where the function must write the output. See the \code{range}
argument from \code{googlesheets4::range_write()} to learn more.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
\code{sqlr} system.

\code{write_query()} create queries for each row of the \code{search} dataset and
write them in the \code{query} column.
}
\examples{
\dontrun{
write_query()}
}
\seealso{
Other Google Sheets functions: 
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{sheet_nrow}()},
\code{\link{write_metadata}()},
\code{\link{write_sheet}()}
}
\concept{Google Sheets functions}
