% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_query.R
\name{write_query}
\alias{write_query}
\title{Write queries in the \code{search} dataset}
\usage{
write_query(range = NULL, package = NULL)
}
\arguments{
\item{range}{(optional) a string indicating the cell range or start of the
target rectangle where the function must write the output. See the \code{range}
argument from \code{googlesheets4::range_write()} to learn more.}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{write_query()} create queries for each row of the \code{search} dataset and
write them in the \code{query} column.
}
\examples{
\dontrun{
write_query()}
}
\seealso{
Other SQLR system functions: 
\code{\link{build_reference}()},
\code{\link{keyword_set}()},
\code{\link{normalize_extdata}()},
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{selection_stats}()},
\code{\link{sheet_nrow}()},
\code{\link{start_sqlr}()},
\code{\link{update_reference}()},
\code{\link{write_metadata}()},
\code{\link{write_reference}()},
\code{\link{write_sheet}()},
\code{\link{write_trial}()}
}
\concept{SQLR system functions}
