% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reference.R
\name{write_reference}
\alias{write_reference}
\title{Write the \code{reference} table to the package and to Google Spreadsheets}
\usage{
write_reference(data, package = gutils:::get_package_name())
}
\arguments{
\item{data}{The \code{data.frame} object to write.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{write_reference()} writes a \code{data.frame} to the reference data file of
the package an to its Google Spreadsheets file.

You must have a \code{sheets} data object with the sheets metadata before running
this function. See \code{\link[=write_metadata]{write_metadata()}} to learn more.
}
\examples{
\dontrun{
write_reference(data.frame(a = 1, b = 1))}
}
\seealso{
Other SQLR system functions: 
\code{\link{build_reference}()},
\code{\link{keyword_set}()},
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{selection_stats}()},
\code{\link{sheet_nrow}()},
\code{\link{start_sqlr}()},
\code{\link{update_reference}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_sheet}()},
\code{\link{write_trial}()}
}
\concept{SQLR system functions}
