% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_trial.R
\name{write_trial}
\alias{write_trial}
\title{Write a trial table to a Google Spreadsheets}
\usage{
write_trial(trial_id, package = NULL, quiet = FALSE)
}
\arguments{
\item{trial_id}{A string indicating the ID of the trial to be written.}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default: \code{NULL}).}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
SQLR system.

\code{write_trial()} writes a \verb{trial_*} table of the Systematic Quantitative
Literature Review (SQLR) system to a Google Spreadsheets.

You must have a \code{sheets} data object with the sheets metadata and an updated
\code{reference} table before running this function. See \code{\link[=write_metadata]{write_metadata()}} and
\code{\link[=update_reference]{update_reference()}} to learn more.
}
\examples{
\dontrun{
## Writing the Trial - DUP table

write_trial("dup")

## Checking the number of rows with the 'reference' table

nrow(dplyr::filter(reference, !is.na(criteria_id)) +
sheet_nrow("trial_nr1") == nrow(reference)}
}
\seealso{
Other SQLR system functions: 
\code{\link{build_reference}()},
\code{\link{keyword_set}()},
\code{\link{normalize_extdata}()},
\code{\link{range_write}()},
\code{\link{read_sheet}()},
\code{\link{sheet_nrow}()},
\code{\link{start_sqlr}()},
\code{\link{update_reference}()},
\code{\link{write_metadata}()},
\code{\link{write_query}()},
\code{\link{write_reference}()},
\code{\link{write_sheet}()}
}
\concept{SQLR system functions}
