% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_trial.R
\name{write_trial}
\alias{write_trial}
\title{Write a trial table to a Google Spreadsheets}
\usage{
write_trial(trial_id, package = rutils:::get_package_name())
}
\arguments{
\item{trial_id}{A string indicating the ID of the trial to be written.}

\item{package}{(optional) a string indicating the package with the database
data. If not assigned, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{CAUTION}: This function must be used only with packages that follow the
\code{sqlr} system.

\code{write_trial()} writes a \verb{trial_*} table of the systematic quantitative
literature review (\code{sqlr}) system to a Google Spreadsheets.

You must have a \code{sheets} data object with the sheets metadata and an updated
\code{reference} table before running this function. See \code{\link[=write_metadata]{write_metadata()}} and
\code{\link[=update_reference]{update_reference()}} to learn more.
}
\examples{
\dontrun{
## Writing the Trial - DUP table

write_trial("dup")

## Checking the number of rows with the 'reference' table

nrow(dplyr::filter(reference, !is.na(criteria_id)) +
sheet_nrow("trial_nr1") == nrow(reference)}
}
\seealso{
Other reference/citation functions: 
\code{\link{build_reference}()},
\code{\link{update_reference}()},
\code{\link{write_reference}()}
}
\concept{reference/citation functions}
